Move toggle logic to seperate trait
